requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pyTelegramBotAPI==4.26.0
requests==2.32.3
urllib3==2.3.0
python-dotenv==1.0.0
Flask==3.0.0
